name conventions & carry heatmap internals to `data` crate
